audio-server/requirements.txt
paho-mqtt
pyttsx3
python-vlc
requests